Switch to swaggers Android template
The Java template generates too large bindings, the android template is more compatible but requires adjustments to the dependencies and the build script. Gradle The generated code exposes properties that GSON can't handle. The problem is patched in the GsonConfiguration: properties exposed by parent classes are skipped. Events are in this version successfully listed on the main screen (Tested on Huawei ALE-L21).