Patch Swagger generator
The code generator does not utilize the feature for inheritance. A new version is compiled from the current head of the git repositories master branch. Now, erroneous reconfiguration of Gson is not required. No fields are duplicated. Adjustments to the build script and the gradle configuration to speed up launch were added.