Rewrite database and create synchronization
The old database implementation was too large to maintain, a new, smaller version is implemented and a set of abstractions is introduced. The background service is retrieving the data in the background and feeding it to the database. There are some issues with the fault tolerance in input and output, which need to be addressed.