Use Volley and remove Roboguice
Roboguice has not been used and not proven useful for the addressed domain, therefore it was removed. Instead of the legacy HTTP API, volley is now introduced. A fix in the swagger code generator now covers string escaping. The app has been tested from cache-free launch and with a delta application, the prior yielded a full update, where the latter had no transmissions due to up-to-date status.