Replace old update mechanism
The previous iteration of the update system relied on a highly complicated system of intents and results. The new one is rather simple. Besides that, cache invalidation was implemented, so that multiple instances of the `Database` can be used across the application without risking invalid results on different ends. A set of refactorings was performed.