Save state on multiple fragments
The following fragments now properly retains view state:

* Dealer list
* Event pager
* Event recycler
* Event detail view

The event pager does not explicitly need to save state, as it does
everything automatically. The problem of the previous implementation
was caused by the recycler not saving state properly.

A lot of calls now rely on the runDelayed method, which is kind-of a
hack but does the trick for now.